更新gf版本为v1.14.0 添加按钮权限 友情链接完善 cms模板选择

+++ app/model/admin/user/user.go
@@ -16,7 +16,7 @@ type StatusReq struct {
 
 //重置用户密码状态参数
 type ResetPwdReq struct {
-	Id       int    `p:"userId" v:"required#用户id不能为空"`
+	Id       uint64 `p:"userId" v:"required#用户id不能为空"`
 	Password string `p:"password" v:"required|password#密码不能为空|密码以字母开头，只能包含字母、数字和下划线，长度在6~18之间"`
 }
